pyproject.toml
[project]
name = "appartment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "selenium>=4.29.0",
    "tqdm>=4.67.1",
    "transformers>=4.49.0",
    "webdriver-manager>=4.0.2",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "accelerate>=0.26.0",
    "bitsandbytes>=0.45.3",
    "qwen-vl-utils[decord]>=0.0.10",
    "pydantic>=2.10.6",
    "ollama>=0.4.7",
    "llmlingua-promptflow>=0.0.1",
    "langchain-ollama>=0.2.3",
    "llmlingua>=0.2.2",
    "streamlit>=1.43.2",
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu124", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu124", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
